requirement.txt
accelerate==0.23.0
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.5
cssselect==1.2.0
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fake-useragent==1.2.1
fastapi==0.103.1
fastapi-pagination==0.12.10
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.4
fsspec==2023.9.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
lit==17.0.1
lxml==4.9.3
Mako==1.2.4
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
orjson==3.9.7
packaging==23.1
parse==1.19.1
Pillow==10.0.1
psutil==5.9.5
psycopg2-binary==2.9.8
pydantic==2.4.1
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
requests-html==0.10.0
safetensors==0.3.3
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
starlette==0.27.0
sympy==1.12
tinysegmenter==0.3
tldextract==3.6.0
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
transformers==4.33.2
triton==2.0.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
w3lib==2.1.2
watchfiles==0.20.0
websockets==10.4
zipp==3.17.0
